hevc: always use ptrdiff_t to represent the stride

+++ libavcodec/hevc.c
@@ -378,7 +378,7 @@ static void hls_residual_coding(HEVCContext *s, int x0, int y0, int log2_trafo_w
 
     const uint8_t *scan_x_cg, *scan_y_cg, *scan_x_off, *scan_y_off;
 
-    int stride = s->frame.linesize[c_idx];
+    ptrdiff_t stride = s->frame.linesize[c_idx];
     int hshift = s->sps->hshift[c_idx];
     int vshift = s->sps->vshift[c_idx];
     uint8_t *dst = &s->frame.data[c_idx][(y0 >> vshift) * stride + (x0 >> hshift)];
