+++ libavcodec/hevcdsp.h
@@ -25,9 +25,9 @@
 
 typedef struct HEVCDSPContext {
     void (*dequant)(int16_t *coeffs, int log2_size, int qp, int bit_depth);
-    void (*transform_4x4_luma_add)(uint8_t *dst, int16_t *coeffs, int stride, int bit_depth);
+    void (*transform_4x4_luma_add)(uint8_t *dst, int16_t *coeffs, ptrdiff_t stride, int bit_depth);
 
-    void (*transform_add[4])(uint8_t *dst, int16_t *coeffs, int _stride, int bit_depth);
+    void (*transform_add[4])(uint8_t *dst, int16_t *coeffs, ptrdiff_t _stride, int bit_depth);
 
 } HEVCDSPContext;
